Arreglar imágenes iteraciones y registrar porcentaje de avance caso uso

+++ Base de Datos/MySQL/XAMPP/createXAMPP.sql
@@ -373,7 +373,7 @@
         IN idFase INT,
         IN idTarea INT,
         IN idAp INT,
-        IN Xminutos INT
+        IN Xminutos DECIMAL(5,1)
     )
     BEGIN
         UPDATE ap_colaborador SET minutos = Xminutos
